+++ SpiffWorkflow/dmn/specs/BusinessRuleTask.py
@@ -1,4 +1,4 @@
-from ...exceptions import WorkflowTaskExecException
+from SpiffWorkflow.bpmn.exceptions import WorkflowTaskExecException
 
 from ...specs import Simple
