+++ tests/SpiffWorkflow/bpmn/ScriptTest.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
 
 import unittest
-from unittest import skip
 
-from SpiffWorkflow.exceptions import WorkflowTaskExecException
+from SpiffWorkflow.bpmn.exceptions import WorkflowTaskExecException
 from SpiffWorkflow.task import TaskState
 from SpiffWorkflow.bpmn.workflow import BpmnWorkflow
 from tests.SpiffWorkflow.bpmn.BpmnWorkflowTestCase import BpmnWorkflowTestCase
